Conflicts:
	examples/glyphs/gauges.py
	examples/glyphs/gears.py

+++ setup.py
@@ -233,23 +233,24 @@ if '--user' in sys.argv:
 else:
     site_packages = getsitepackages()[0]
 
-bokeh_path = join(site_packages, "bokeh")
-if exists(bokeh_path) and isdir(bokeh_path):
-    prompt = "\nFound existing bokeh install: %s\nRemove it? [y|N] " % bokeh_path
-    if sys.version_info[0] < 3:
-        val = raw_input(prompt)
-    else:
-        val = input(prompt)
-    if val == "y":
-        print ("Removing old bokeh install...")
-        try:
-            shutil.rmtree(bokeh_path)
-            print ("Removed.")
-        except (IOError, OSError):
-            print ("Unable to remove old bokeh at: %s" % bokeh_path)
-            sys.exit(-1)
-    else:
-        print ("Not removing old bokeh install")
+if 'develop' in sys.argv:
+    bokeh_path = join(site_packages, "bokeh")
+    if exists(bokeh_path) and isdir(bokeh_path):
+        prompt = "\nFound existing bokeh install: %s\nRemove it? [y|N] " % bokeh_path
+        if sys.version_info[0] < 3:
+            val = raw_input(prompt)
+        else:
+            val = input(prompt)
+        if val == "y":
+            print ("Removing old bokeh install...")
+            try:
+                shutil.rmtree(bokeh_path)
+                print ("Removed.")
+            except (IOError, OSError):
+                print ("Unable to remove old bokeh at: %s" % bokeh_path)
+                sys.exit(-1)
+        else:
+            print ("Not removing old bokeh install")
 
 path_file = join(site_packages, "bokeh.pth")
 path = abspath(dirname(__file__))
